backend/requirements.txt
amqp==5.3.1
asgiref==3.11.0
billiard==4.2.4
celery==5.3.6
chardet==5.2.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.0
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
kombu==5.6.2
packaging==26.0
pillow==10.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
PyJWT==2.11.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-magic==0.4.27
pytz==2025.2
qrcode==7.4.2
redis==5.0.1
reportlab==4.0.9
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
vine==5.1.0
wcwidth==0.5.3
